% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_grid.R
\name{plot_grid}
\alias{plot_grid}
\title{plot_grid}
\usage{
plot_grid(indicator_array, lower_bds, upper_bds)
}
\arguments{
\item{indicator_array}{Indicator array generated using confidence_grid}

\item{lower_bds}{Vector containing the lower bounds for the search space.}

\item{upper_bds}{Vector containing the upper bounds for the search space.}
}
\value{
A grid plot showing the confidence regions
}
\description{
projects the indicator array generated by confidence_grid down to 2d for visualization
}
\examples{
# Regular normal
set.seed(123)
n <- 100 # sample size
R <- 200 # Repro sample size
alpha <- .05 # significance level
tol <- 1e-4 # tolerance for the confidence set
s_obs <- c(1.12, 0.67) # the observed sample mean
seeds <- matrix(rnorm(R * (n + 2)), nrow = R, ncol = n + 2) # pre-generated seeds

# this function computes the repro statistics given the seeds and the parameter
s_sample <- function(seeds, theta) {
  # generate the raw data points
  raw_data <- theta[1] + sqrt(theta[2]) * seeds[, 1:n]

  # compute the regular statistics
  s_mean <- apply(raw_data, 1, mean)
  s_var <- apply(raw_data, 1, var)

  return(cbind(s_mean, s_var))
}

lower_bds <- c(-5, 0.01) # lower bounds for the parameter search region
upper_bds <- c(5, 5) # upper bounds for the parameter search region

res = 20  # resolution of the grid
result <- confidence_grid(alpha, lower_bds, upper_bds, seeds, s_sample, s_obs, tol, res)

ind_arr <- result$ind_array
lower_bds <- result$updated_lower_bds
upper_bds <- result$updated_upper_bds
print(ind_arr)
plot_grid(ind_arr, lower_bds, upper_bds)
print(lower_bds)
print(upper_bds)

}
