% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_grid.R
\name{confidence_grid}
\alias{confidence_grid}
\title{confidence_grid}
\usage{
confidence_grid(
  alpha,
  lower_bds,
  upper_bds,
  seeds,
  G,
  s_obs,
  tol,
  res,
  t_init = NULL,
  T_stat = ma_depth
)
}
\arguments{
\item{alpha}{Significance level.}

\item{lower_bds}{Vector containing the lower bounds for the search space.}

\item{upper_bds}{Vector containing the upper bounds for the search space}

\item{seeds}{Seeds containing all the randomness.}

\item{G}{Data generating process.}

\item{s_obs}{Observed statistic.}

\item{tol}{tolerance of the confidence interval.}

\item{res}{resolution of the grid}

\item{t_init}{Starting point for the initial search.}

\item{T_stat}{A distance metric, default to Mahalanobis distance.}
}
\value{
An array
}
\description{
Plots the confidence grid
}
\examples{
# example1
example1
}
