% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_grid.R
\name{confidence_grid}
\alias{confidence_grid}
\title{confidence_grid}
\usage{
confidence_grid(
  alpha,
  lower_bds,
  upper_bds,
  seeds,
  G,
  s_obs,
  tol,
  res,
  t_init = NULL,
  T_stat = ma_depth
)
}
\arguments{
\item{alpha}{Significance level.}

\item{lower_bds}{Vector containing the lower bounds for the search space.}

\item{upper_bds}{Vector containing the upper bounds for the search space}

\item{seeds}{Seeds containing all the randomness.}

\item{G}{Data generating process.}

\item{s_obs}{Observed statistic.}

\item{tol}{tolerance of the confidence interval.}

\item{res}{resolution of the grid}

\item{t_init}{Starting point for the initial search.}

\item{T_stat}{A distance metric, default to Mahalanobis distance.}
}
\value{
A list containing ind_array, updated_lower_bds, and updated_upper_bds.
}
\description{
returns the indicator array
}
\examples{
### Note that the examples may take a few seconds to run.
### Regular normal
set.seed(123)
n <- 50 # sample size
R <- 200 # Repro sample size
alpha <- .05 # significance level
tol <- 1e-2 # tolerance for the confidence set
s_obs <- c(1.12, 0.67) # the observed sample mean
seeds <- matrix(rnorm(R * (n + 2)), nrow = R, ncol = n + 2) # pre-generated seeds

# this function computes the repro statistics given the seeds and the parameter
s_sample <- function(seeds, theta) {
  # generate the raw data points
  raw_data <- theta[1] + sqrt(theta[2]) * seeds[, 1:n]

  # compute the regular statistics
  s_mean <- apply(raw_data, 1, mean)
  s_var <- apply(raw_data, 1, var)

  return(cbind(s_mean, s_var))
}

lower_bds <- c(0.5, 0.3) # lower bounds for the parameter search region
upper_bds <- c(1.5, 1.3) # upper bounds for the parameter search region

res = 20  # resolution of the grid
result <- confidence_grid(alpha, lower_bds, upper_bds, seeds, s_sample, s_obs, tol, res)
print(result$ind_array)
print(result$search_lower_bds)
print(result$search_upper_bds)

}
