% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CI_function.R
\name{get_CI}
\alias{get_CI}
\title{get_CI}
\usage{
get_CI(
  alpha,
  lower_bds,
  upper_bds,
  j,
  seeds,
  G,
  s_obs,
  tol,
  t_init = NULL,
  T_stat = ma_depth
)
}
\arguments{
\item{alpha}{Significance level.}

\item{lower_bds}{Vector containing the lower bounds for the search space.}

\item{upper_bds}{Vector containing the upper bounds for the search space}

\item{j}{Index indicating the parameter of interest.}

\item{seeds}{Seeds containing all the randomness.}

\item{G}{Data generating process.}

\item{s_obs}{Observed statistic.}

\item{tol}{tolerance of the confidence interval.}

\item{t_init}{Starting point for the initial search.}

\item{T_stat}{A distance metric.}
}
\value{
A vector containing the lower and upper boundary of the confidence interval. In the case when no point is accepted in the search space, return NULL.
}
\description{
Given the observed statistic, this function computes a confidence interval given that the data generating process is known using the supplied random seeds.
}
\examples{
# DP Normal
blahblahblah
}
