% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_value_function.R
\name{p_value}
\alias{p_value}
\title{p_value}
\usage{
p_value(
  lower_bds,
  upper_bds,
  seeds,
  G,
  s_obs,
  t_init = NULL,
  T_stat = ma_depth,
  print_info = FALSE,
  check_input = TRUE
)
}
\arguments{
\item{lower_bds}{Vector containing the lower bounds for the search space.}

\item{upper_bds}{Vector containing the upper bounds for the search space}

\item{seeds}{Seeds containing all the randomness.}

\item{G}{Data generating function.}

\item{s_obs}{Observed statistic.}

\item{t_init}{Starting point for the initial search.}

\item{T_stat}{A distance metric (default to ma_depth which measures the Mahalanobis distance).}

\item{print_info}{Whether or not to print the optim information}

\item{check_input}{Whether or not to run checks on the function inputs.}
}
\value{
This function returns a list containing the minimum p value (p_val) within the search region and the parameter corresponding to it (theta_hat).
}
\description{
Given the observed statistic and the given seeds, this function finds the largest p_value and the corresponding parameter
}
\examples{
### Regular Normal
set.seed(123)
n <- 50 # sample size
R <- 200 # Repro sample size
s_obs <- c(1.12, 0.67) # the observed sample mean and variance
seeds <- matrix(rnorm(R * (n + 2)), nrow = R, ncol = n + 2) # pre-generated seeds

# this function computes the repro statistics given the seeds and the parameter
s_sample <- function(seeds, theta) {
  # generate the raw data points
  raw_data <- theta[1] + sqrt(theta[2]) * seeds[, 1:n]

  # compute the regular statistics
  s_mean <- apply(raw_data, 1, mean)
  s_var <- apply(raw_data, 1, var)

  return(cbind(s_mean, s_var))
}

lower_bds <- c(-5, 0.01) # lower bounds for the parameter search region
upper_bds <- c(5, 5) # upper bounds for the parameter search region

result <- p_value(lower_bds, upper_bds, seeds, s_sample, s_obs)
print(result$p_val) # the largest p_value found
print(result$theta_hat) # the parameter corresponding to the largest p value


}
